astroshield-integration-package/src/asttroshield/udl_integration/requirements.txt
requests>=2.25.0
confluent-kafka>=2.0.0
python-dateutil>=2.8.1
pyyaml>=6.0
python-dotenv>=0.15.0
prometheus-client>=0.10.0
jsonschema>=4.0.0
tenacity>=8.0.0
urllib3>=1.26.0
cachetools>=4.2.0
typing-extensions>=4.0.0
pytest>=7.0.0
mock>=5.0.0
kafka-python>=2.0.2
threading2>=0.3.1 